FastFlow-1.1.0 release, see the Changelog.txt file for full details.

+++ ff/farm.hpp
@@ -151,7 +151,7 @@ public:
             workers[i] = w[i];
             workers[i]->set_id(i);
         }
-        nworkers+=w.size();
+        nworkers+= (unsigned int) w.size();
         return 0;
     }
